[16.0][REF] kris_project: merge kris_project_exception into kris_project
Move base.exception inheritance and the state-transition action methods
(action_confirm, action_cancel, action_draft) from the split-off
kris_project_exception.py into the main kris_project.py so the model
definition and its actions live in one file.
